requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
pydantic<2.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
python-docx>=1.1.0
numpy>=1.24.0
requests>=2.31.0
streamlit>=1.30.0
openai>=1.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
